# Add benchmarks, optimize, and simplify in new netstack