--- a/internal/iocount/nop.go
+++ b/internal/iocount/nop.go
@@ -1,0 +1,16 @@
+package iocount
+
+import "io"
+
+type NopWriteCloser struct {
+	writer io.Writer
+}
+
+func NewNopWriteCloser(w io.Writer) *NopWriteCloser {
+	return &NopWriteCloser{
+		writer: w,
+	}
+}
+
+func (n *NopWriteCloser) Write(b []byte) (int, error) { return n.writer.Write(b) }
+func (n *NopWriteCloser) Close() error                { return nil }