--- a/internal/wireguard/device_interface.go
+++ b/internal/wireguard/device_interface.go
@@ -6,14 +6,13 @@ import (
 	"log/slog"
 	"net/netip"
 
+	"github.com/johnstarich/zfs-sync-operator/internal/netstack"
 	"golang.zx2c4.com/wireguard/conn"
 	wgdevice "golang.zx2c4.com/wireguard/device"
-	"golang.zx2c4.com/wireguard/tun/netstack"
 )
 
 // DeviceInterface starts, runs, and stops a WireGuard device's interface in userspace.
 type DeviceInterface struct {
-	dnsAddresses []netip.Addr
 	done         chan struct{}
 	localAddress netip.Addr
 	logger       *slog.Logger
@@ -21,16 +20,8 @@ type DeviceInterface struct {
 }
 
 // newInterface returns a new [DeviceInterface] for the given interface name
-func newInterface(logger *slog.Logger, localAddress netip.Addr, dnsAddresses []netip.Addr) *DeviceInterface {
-	if len(dnsAddresses) == 0 {
-		// CloudFlare DNS
-		dnsAddresses = []netip.Addr{
-			netip.MustParseAddr("1.0.0.1"),
-			netip.MustParseAddr("1.1.1.1"),
-		}
-	}
+func newInterface(logger *slog.Logger, localAddress netip.Addr) *DeviceInterface {
 	return &DeviceInterface{
-		dnsAddresses: dnsAddresses,
 		done:         make(chan struct{}),
 		localAddress: localAddress,
 		logger:       logger,
@@ -40,14 +31,14 @@ func newInterface(logger *slog.Logger, localAddress netip.Addr, dnsAddresses []n
 // Start starts and runs the interface until the context is canceled.
 // Use [Wait] to view any errors encountered while running.
 func (i *DeviceInterface) Start(ctx context.Context) (*wgdevice.Device, *netstack.Net, error) {
-	tunDevice, tunNet, err := netstack.CreateNetTUN([]netip.Addr{i.localAddress}, i.dnsAddresses, wgdevice.DefaultMTU)
+	tunNet, err := netstack.New([]netip.Addr{i.localAddress}, wgdevice.DefaultMTU)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	const maxDeviceErrors = 2
 	logger := newLastErrorsLogger(maxDeviceErrors, "wireguard device")
-	device := wgdevice.NewDevice(tunDevice, conn.NewDefaultBind(), &wgdevice.Logger{
+	device := wgdevice.NewDevice(tunNet, conn.NewDefaultBind(), &wgdevice.Logger{
 		Verbosef: func(format string, args ...any) {
 			i.logger.Debug(fmt.Sprintf(format, args...))
 		},