Add benchmarks, optimize, and simplify in new netstack

Full changes:
* Update Backup status instantly, then on interval
* Add benchmark for Backup transfer rate over SSH
* Include enhancements to test SSH server for streaming data without large allocations
* Add WireGuard benchmark
* Upgrade wireguard-go to latest
* Drop ignore for dataset exists, indicative of a bad transfer state. Might need to reintroduce a fix for this, but leaving for manual resolution for now.
* Create custom WireGuard netstack for optimizations, remove DNS support for now
* Reduce net allocations by reading directly from packet into byte slice
* Add netstack stats to log, includes error information
* Ignore context cancelation in WireGuard interface waiter, its an expected result
* Extract io counters into new package

Benchmark results before custom netstack (note: noisy data):

    goos: linux
    goarch: amd64
    pkg: internal/backup
    cpu: 13th Gen Intel(R) Core(TM) i7-1370P
    BenchmarkSendSpeed/SSH_to_SSH-20                5371           2055948 ns/op         510.02 MB/s
    BenchmarkSendSpeed/SSH_to_WireGuard-20          640          22243412 ns/op          47.14 MB/s
    BenchmarkSendSpeed/WireGuard_to_SSH-20          668          26171884 ns/op          40.06 MB/s
    BenchmarkSendSpeed/WireGuard_to_WireGuard-20    100         137669484 ns/op           7.62 MB/s

And after:

    goos: linux
    goarch: amd64
    pkg: internal/backup
    cpu: 13th Gen Intel(R) Core(TM) i7-1370P
    BenchmarkSendSpeed/SSH_to_SSH-20              6061           1685906 ns/op         621.97 MB/s
    BenchmarkSendSpeed/SSH_to_WireGuard-20        1038          11336430 ns/op          92.50 MB/s
    BenchmarkSendSpeed/WireGuard_to_SSH-20        1198          10154468 ns/op         103.26 MB/s
    BenchmarkSendSpeed/WireGuard_to_WireGuard-20  2016           8318924 ns/op         126.05 MB/s